代码优化

+++ src/main/java/top/yinzsw/blog/model/dto/EmailCodeDTO.java
@@ -29,5 +29,5 @@ public class EmailCodeDTO {
     /**
      * 有效期分钟数
      */
-    private int timeoutMinutes;
+    private long timeoutMinutes;
 }
